+++ src/Weapons/Shotgun.cpp
@@ -38,16 +38,16 @@ bool Shotgun::fire(GameWorld* world, WorldEntity* entity)
         int openAngle = 24;
         for (int i(0); i<12; ++i)
         {
-            float angle = rand()%openAngle-openAngle/2;
-            float speed = (rand()%50+125)/100.f;
+            float angle = static_cast<float>(rand()%openAngle-openAngle/2);
+            float speed = (rand()%50+125)/100.0f;
             Bullet* newBullet = Bullet::add(angle*DEGRAD, speed*CELL_SIZE, 15, 30);
             newBullet->init(entity->getCoord(), entityAngle);
             world->addEntity(newBullet);
 
             Vec2 bulletVel(newBullet->getV());
             Vec2 smokePos = newBullet->getCoord()+bulletVel*1.25f;
-            float v(rand()%50/1000.0f+0.1);
-            world->addEntity(Smoke::add(smokePos, bulletVel*v, 0.05, 75));
+            float v(rand()%50/1000.0f+0.1f);
+            world->addEntity(Smoke::add(smokePos, bulletVel*v, 0.05f, 75));
 
             Vec2 firePos(newBullet->getCoord()+bulletVel*_fireDist);
             world->addEntity(Fire::add(firePos, entityAngle+PIS2));
